refactor(slice-45): split scripts into ci/docker/release/security themes
Complete hotspot #5 and refresh the theme map to IMPLEMENTED paths;
keep thin flat shims for one minor so old invocations still work (DECISIONS #159).
