+++ docs/plan/DECISIONS.md
@@ -143,7 +143,7 @@
 | 137 | 2026-07-27 | 44/45 | nw-review remediations applied to SLICE-44 + SLICE-45 stubs | Locked gate scripts (`test:coverage` local / `test:ci` CI); usage metrics; §3 vs COMPLETE clarified; Slice 45 Reuse Analysis + re-export lifecycle + blast-radius `rg` + quantified GWT. PO Elevator Pitch already present. Superseding verdicts: 44 DoR APPROVED for execution; 45 architect APPROVED | Leave NEEDS_REVISION / CONDITIONALLY APPROVED without stub updates |
 | 138 | 2026-07-27 | 44 | Frontend coverage floor locked then ratcheted: Before-Check baseline lines **50.18%** / branches **47.74%** / functions **45.38%** / statements **48.48%**; after Should tests measured ≈64.75/58.22/61.15/62.63 — thresholds set to **64/58/61/62**. Mutation testing waived (CI-hygiene; no non-trivial new pure logic) | Honest floor from re-measure; ratchet when tests raise coverage so gates catch regressions | Keep baseline 50% floor after Should tests / invent aspirational 100% |
 | 141 | 2026-07-27 | coverage | Fair metric breakdown: FE **95/90/95/95**; BE policy **92/85/n/a/90** with enforced `fail_under=90` (combined Cover) | Interim after #140. **Superseded by #142** (BE raised + JSON checker) | Keep flat FE 90 / invent BE branch≥90 without raising measured branch |
-| 142 | 2026-07-27 | coverage | Shared product floors **95/90/95/95** for FE; BE **95/90/n/a/95** (functions n/a on coverage.py) | FE Vitest-enforced. BE: gap tests (`test_coverage_floor_gaps.py`) + `fail_under` **90→95** + `scripts/check_backend_coverage_floors.py` (stmts/br/lines from JSON; floors in `[tool.rag_params_finder.coverage_thresholds]`). Wired in quality-gates / pre-push / CI. Measured stmts≈98.6 / br≈95.2 / TOTAL≈97.7. Supersedes #141 soft BE 92/85 | Keep BE fail_under=90 / aim-only without JSON checker |
+| 142 | 2026-07-27 | coverage | Shared product floors **95/90/95/95** for FE; BE **95/90/n/a/95** (functions n/a on coverage.py) | FE Vitest-enforced. BE: gap tests (`test_coverage_floor_gaps.py`) + `fail_under` **90→95** + `scripts/ci/check_backend_coverage_floors.py` (stmts/br/lines from JSON; floors in `[tool.rag_params_finder.coverage_thresholds]`). Wired in quality-gates / pre-push / CI. Measured stmts≈98.6 / br≈95.2 / TOTAL≈97.7. Supersedes #141 soft BE 92/85 | Keep BE fail_under=90 / aim-only without JSON checker |
 | 140 | 2026-07-27 | coverage | Uniform product coverage floor **≥90%** for backend and frontend | Backend unit scoped gate `--cov-fail-under` / `fail_under` **80→90** (measured TOTAL ≈93.2%). Frontend Vitest thresholds briefly **90/90/90/90**. Postgres `retriever_postgres` CI module gate stays **≥95%** (specialized). **FE quartet superseded by #141**; BE fail_under=90 retained until #142 | Keep asymmetric FE 95 / BE 80 |
 | 139 | 2026-07-27 | 44 | Phase B reopen: FE coverage gate → **≥95%** stmts/funcs/lines + **≥90%** branches (`all: true` on `src/**`; exclude tests/setup/main/types) | User chose option 1 + reopen 44 (not 44B). Surpasses backend `--cov-fail-under=80`. Supersedes #132/#138 “100% Won’t / 64% floor” for product thresholds. **Superseded for product floors by #140** | Keep 64% floor / invent literal 100% without exclusions / new Slice 44B |
 | 143 | 2026-07-27 | health-check | Gap 8: Specification coverage gates inserted into SLICE-45 After-Checks | Active PLANNED stub lacked §12 coverage / branch / mutation After-Check lines; user confirmed (y); product floors #142 retained (not whole-tree 100%) | SKIPPED — leave Gap 8 open on Slice 45 |
@@ -159,3 +159,4 @@
 | 153 | 2026-07-28 | 45 | Fat-surface slim: `experiments_lifecycle.py` + `cli/display.py`; keep routers/commands as façades | Execution order step 3 after db; pre→post lines experiments 553→388, cli/main 519→246 | Defer slim until FE primitives / leave god modules |
 | 157 | 2026-07-28 | 45 | Split `frontend/src/components/` into `screens/`, `chrome/`, `experiment/`, `stats/` (+ co-located tests); nest `experimentDetail` under `experiment/` | Step 7 after SLAP; App + mocks retargeted; FE floors still green | Leave flat components/ / move without co-located tests |
 | 158 | 2026-07-28 | 45 | FE shared test helpers under `frontend/src/test/helpers/` (`experiments`, `vectorDbStats`, `explore`, `experimentDetail`); screen + stats suites import builders | Step 8 After-Check; `src/test/**` already coverage-excluded; 261 tests / floors unchanged | Leave duplicated builders in mega-suites / invent factories only for new tests |
+| 159 | 2026-07-28 | 45 | Land `scripts/{ci,docker,release,security}/`; keep `lib/`; thin shims at old `scripts/<name>` for one minor | Completes hotspot #5 Could; living docs/CI/pre-commit use canonical paths; historical slice specs left as snapshots | Big-bang delete old paths without shims / leave scripts flat |
